refactor(ui): dedupe chat tooltip icon buttons into TooltipIconButton
